refactor: split opencode adapter normalizers
